add reset approval PR & PO

+++ app/Http/Controllers/Transaksi/CancelApprovePoController.php
@@ -18,7 +18,9 @@ class CancelApprovePoController extends Controller
             $params = $request->params;        
             $whereClause = $params['sac'];
         }
-        $query = DB::table('t_po01')
+        $query = DB::table('v_rpo')
+                 ->select('id', 'ponum', 'podat', 'vendor_name', 'note')
+                 ->distinct()
                  ->orderBy('id');
         return DataTables::queryBuilder($query)
         ->toJson();
@@ -29,31 +31,42 @@ class CancelApprovePoController extends Controller
         try{
             $wodata = DB::table('t_po01')->where('id', $id)->first();
             if($wodata){
-                DB::table('t_po01')->where('id', $id)->update([
-                    'approvestat'   => 'N'
-                ]);
 
-                $firstApproval = DB::table('t_po_approval')
-                        ->where('ponum', $wodata->ponum)
-                        ->orderBy('approver_level', 'ASC')
-                        ->first();
-
-                DB::table('t_po_approval')->where('ponum', $wodata->ponum)->update([
-                    'approval_status' => 'N',
-                    'is_active'       => 'N'
-                ]);
-
-                DB::table('t_po_approval')->where('ponum', $wodata->ponum)
-                ->where('approver_level', $firstApproval->approver_level)
-                ->update([
-                    'is_active' => 'Y'
-                ]);
-                DB::commit();
-                
-                $result = array(
-                    'msgtype' => '200',
-                    'message' => 'Approval PO '. $wodata->ponum . ' berhasil direset'
-                );
+                $checkGR = DB::table('t_po02')
+                    ->where('ponum', $wodata->ponum)
+                    ->where('grqty', '>', 0)->first();
+                if($checkGR){
+                    $result = array(
+                        'msgtype' => '500',
+                        'message' => 'PO '. $wodata->ponum . ' sudah ada proses penerimaan barang'
+                    );
+                }else{
+                    DB::table('t_po01')->where('id', $id)->update([
+                        'approvestat'   => 'N'
+                    ]);
+    
+                    $firstApproval = DB::table('t_po_approval')
+                            ->where('ponum', $wodata->ponum)
+                            ->orderBy('approver_level', 'ASC')
+                            ->first();
+    
+                    DB::table('t_po_approval')->where('ponum', $wodata->ponum)->update([
+                        'approval_status' => 'N',
+                        'is_active'       => 'N'
+                    ]);
+    
+                    DB::table('t_po_approval')->where('ponum', $wodata->ponum)
+                    ->where('approver_level', $firstApproval->approver_level)
+                    ->update([
+                        'is_active' => 'Y'
+                    ]);
+                    DB::commit();
+                    
+                    $result = array(
+                        'msgtype' => '200',
+                        'message' => 'Approval PO '. $wodata->ponum . ' berhasil direset'
+                    );
+                }
             }else{
                 $result = array(
                     'msgtype' => '500',
@@ -77,14 +90,25 @@ class CancelApprovePoController extends Controller
             $wodata = DB::table('t_po01')->where('id', $id)->first();
             if($wodata){
                 DB::table('t_po01')->where('id', $id)->delete();
-                DB::table('t_po02')->where('ponum', $wodata->ponum)->delete();
-                DB::table('t_po_approval')->where('ponum', $wodata->ponum)->delete();
-                DB::commit();
-
-                $result = array(
-                    'msgtype' => '200',
-                    'message' => 'PO '. $wodata->ponum . ' berhasil dihapus'
-                );
+                $checkGR = DB::table('t_po02')
+                    ->where('ponum', $wodata->ponum)
+                    ->where('grqty', '>', 0)->first();
+                if($checkGR){
+                    $result = array(
+                        'msgtype' => '500',
+                        'message' => 'PO '. $wodata->ponum . ' sudah ada proses penerimaan barang'
+                    );
+                }else{
+                    DB::table('t_po02')->where('ponum', $wodata->ponum)->delete();
+                    DB::table('t_po03')->where('ponum', $wodata->ponum)->delete();
+                    DB::table('t_po_approval')->where('ponum', $wodata->ponum)->delete();
+                    DB::commit();
+    
+                    $result = array(
+                        'msgtype' => '200',
+                        'message' => 'PO '. $wodata->ponum . ' berhasil dihapus'
+                    );
+                } 
             }else{
                 $result = array(
                     'msgtype' => '500',
